requirements.txt
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
jinja2>=3.1.0
Pillow>=9.0.0
numpy>=1.24.0
